+++ moco-runner/src/main/java/com/github/dreamhead/moco/parser/model/GlobalSetting.java
@@ -11,7 +11,6 @@ public class GlobalSetting {
     private String fileRoot;
     private String env;
     private ResponseSetting response;
-    private HttpsSetting https;
 
     public String getInclude() {
         return join(fileRoot, include);
@@ -31,9 +30,5 @@ public class GlobalSetting {
 
     public ResponseSetting getResponse() {
         return response;
-    }
-
-    public HttpsSetting getHttps() {
-        return https;
     }
 }