+++ moco-doc/ReleaseNotes.md
@@ -10,6 +10,8 @@
 ** endsWith
 ** contain
 ** exist
+* between for verifier
+* Added multiple monitors for server creation.
 
 ### Template
 * change template variables from Object to String.