+++ moco-doc/ReleaseNotes.md
@@ -1,5 +1,4 @@
-# Release
-
+# Release 0.9.2 (1-Jul-2014)
 ## Usage
 * (Beta) HTTPS support